# openapi.yaml
openapi: 3.0.3
info:
  title: NOC Chamados API
  description: |
    API REST para sistema de gestão de chamados de NOC (Network Operations Center).
    
    ## Autenticação
    A API usa autenticação JWT via HTTPOnly Cookie. Faça login primeiro usando `/auth/login`.
    
    ## Rate Limiting
    - Login: 5 tentativas por 15 minutos (por IP)
    
    ## Recursos
    - [Documentação Interativa](/api-docs)
    - [Guia Swagger](./SWAGGER_GUIDE.md)
    - [Coleção Postman](./NOC_Chamados.postman_collection.json)
    
  version: 1.2.3
  contact:
    name: Suporte NOC Chamados
  license:
    name: Proprietário
servers:
  - url: https://chamados.xmov.com.br/api
    description: Produção
  - url: http://localhost:3000/api
    description: Desenvolvimento Local
...
paths:
...
  /auth/login:
    post:
      tags:
        - Autenticação
      summary: Login
      description: |
        Autentica um usuário e retorna um token JWT em cookie HTTPOnly.
        
        **Rate Limit:** 5 tentativas por 15 minutos (por IP)
      operationId: login
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - email
                - senha
              properties:
                email:
                  type: string
                  format: email
                  example: admin@xmov.com.br
                senha:
                  type: string
                  format: password
                  minLength: 8
                  example: Admin@123
      responses:
        '200':
          description: Login bem-sucedido
          headers:
            Set-Cookie:
              schema:
                type: string
                example: token=eyJhbGc...; HttpOnly; Secure; SameSite=Strict
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  nome:
                    type: string
                    example: Administrador
                  perfil:
                    type: string
                    enum: [operador, supervisor]
                    example: supervisor
        '401':
          description: Email ou senha incorretos
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '429':
          description: Muitas tentativas de login
          headers:
            Retry-After:
              schema:
                type: integer
                example: 900
            X-RateLimit-Limit:
              schema:
                type: integer
                example: 5
            X-RateLimit-Remaining:
              schema:
                type: integer
                example: 0
            X-RateLimit-Reset:
              schema:
                type: string
                format: date-time
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
...
  /chamados/{id}/fechar:
    post:
      tags:
        - Chamados
      summary: Fechar Chamado
      description: Fecha um chamado com informações de resolução
      operationId: fecharChamado
      security:
        - cookieAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/FecharChamado'
      responses:
        '200':
          description: Chamado fechado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Chamado'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
  /chamados/{id}/historico:
    get:
      tags:
        - Chamados
      summary: Histórico do Chamado
      description: Lista o histórico de um chamado
      operationId: obterHistoricoChamado
      security:
        - cookieAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: Histórico do chamado
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/ChamadoHistorico'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
...
  /chamados/{id}/anexos:
    get:
      tags:
        - Anexos
      summary: Listar Anexos
      description: Lista anexos de um chamado
      operationId: listarAnexos
      security:
        - cookieAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: Lista de anexos
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Anexo'

    post:
      tags:
        - Anexos
      summary: Upload de Anexo
      description: |
        Faz upload de um anexo para o chamado.
        
        **Tipos permitidos:** JPEG, PNG, GIF, WebP, PDF, TXT, CSV, DOC, DOCX, XLS, XLSX
        
        **Tamanho máximo:** 10MB
      operationId: uploadAnexo
      security:
        - cookieAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              required:
                - file
              properties:
                file:
                  type: string
                  format: binary
      responses:
        '201':
          description: Anexo criado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Anexo'
  /chamados/{id}/anexos/{anexoId}:
    delete:
      tags:
        - Anexos
      summary: Deletar Anexo
      description: Remove um anexo
      operationId: deletarAnexo
      security:
        - cookieAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
        - name: anexoId
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: Anexo deletado
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string

  # ==============================================
  # LINKS
  # ==============================================
...
components:
  securitySchemes:
    cookieAuth:
      type: apiKey
      in: cookie
      name: token
      description: JWT token armazenado em HTTPOnly cookie
  responses:
    Unauthorized:
      description: Não autenticado
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            error: Não autorizado
...
    BadRequest:
      description: Dados inválidos
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            error: Dados inválidos
    NotFound:
      description: Recurso não encontrado
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            error: Recurso não encontrado
  schemas:
    Error:
      type: object
      properties:
        error:
          type: string
          example: Mensagem de erro
...
    Chamado:
      type: object
      properties:
        id:
          type: integer
          example: 1
        numero:
          type: string
          example: '2025.12345'
        status:
          type: string
          enum: [aberto, em_atendimento, aguardando_operadora, fechado]
          example: aberto
        impacto:
          type: string
          enum: [critico, alto, medio, baixo]
          example: critico
        tipoChamado:
          type: string
          enum: [link, transporte]
          example: link
        descricaoProblema:
          type: string
          example: Link fora do ar
        protocoloOperadora:
          type: string
          nullable: true
          example: INC123456
        dataDeteccao:
          type: string
          format: date-time
          example: '2025-12-15T10:00:00.000Z'
        dataAbertura:
          type: string
          format: date-time
          example: '2025-12-15T10:05:00.000Z'
        dataNormalizacao:
          type: string
          format: date-time
          nullable: true
          example: '2025-12-15T14:30:00.000Z'
        dataFechamento:
          type: string
          format: date-time
          nullable: true
        link:
          $ref: '#/components/schemas/Link'
          nullable: true
        transporte:
          $ref: '#/components/schemas/Transporte'
          nullable: true
        tipoFalha:
          $ref: '#/components/schemas/TipoFalha'
        abertoPor:
          $ref: '#/components/schemas/UsuarioSimples'
        fechadoPor:
          $ref: '#/components/schemas/UsuarioSimples'
          nullable: true
...
    FecharChamado:
      type: object
      required:
        - dataResolucao
      properties:
        dataResolucao:
          type: string
          format: date-time
          example: '2025-12-15T15:00:00.000Z'
        semConfirmacaoOperadora:
          type: boolean
          default: false
          example: false
        observacao:
          type: string
          example: Serviço normalizado
    ChamadoHistorico:
      type: object
      properties:
        id:
          type: integer
        tipoAcao:
          type: string
          example: atualizacao
        descricao:
          type: string
          example: Status alterado para Em atendimento
        criadoEm:
          type: string
          format: date-time
        usuario:
          $ref: '#/components/schemas/UsuarioSimples'
...
    Anexo:
      type: object
      properties:
        id:
          type: integer
        nome:
          type: string
          example: evidencia.jpg
        tipo:
          type: string
          example: image/jpeg
        tamanho:
          type: integer
          example: 524288
        url:
          type: string
          example: /uploads/chamados/1/evidencia.jpg
        criadoEm:
          type: string
          format: date-time
    Link:
      type: object
      properties:
        id:
          type: integer
          example: 1
        designador:
          type: string
          example: LINK-001
        tipoServico:
          type: string
          enum: [LINK IP, TRANSPORTE]
          example: LINK IP
        capacidade:
          type: string
          example: 10G
        diaVencimento:
          type: integer
          nullable: true
          minimum: 1
          maximum: 31
          example: 25
        ativo:
          type: boolean
          example: true
        enderecoPontaA:
          type: string
          nullable: true
        enderecoPontaB:
          type: string
          nullable: true
        observacoes:
          type: string
          nullable: true
        operadora:
          $ref: '#/components/schemas/Operadora'
        pop:
          $ref: '#/components/schemas/Pop'
        transporte:
          $ref: '#/components/schemas/Transporte'
          nullable: true
...
    Transporte:
      type: object
      properties:
        id:
          type: integer
          example: 1
        nome:
          type: string
          example: TRANSPORTE-001
        fornecedor:
          type: string
          nullable: true
          example: Fornecedor A
        origem:
          type: string
          nullable: true
          example: São Paulo
        destino:
          type: string
          nullable: true
          example: Rio de Janeiro
        ativo:
          type: boolean
          example: true
        observacoes:
          type: string
          nullable: true
        _count:
          type: object
          properties:
            links:
              type: integer
            chamados:
              type: integer
...
    Operadora:
      type: object
      properties:
        id:
          type: integer
          example: 1
        nome:
          type: string
          example: Operadora A
        telefoneSuporte:
          type: string
          nullable: true
          example: '+55 11 1234-5678'
        emailSuporte:
          type: string
          format: email
          nullable: true
          example: suporte@operadora.com
        portalUrl:
          type: string
          format: uri
          nullable: true
          example: https://portal.operadora.com
        slaHoras:
          type: integer
          nullable: true
          example: 4
        ativo:
          type: boolean
          example: true
        observacoes:
          type: string
          nullable: true
...
    Pop:
      type: object
      properties:
        id:
          type: integer
          example: 1
        codigo:
          type: string
          example: SP01
        nome:
          type: string
          nullable: true
          example: POP São Paulo
        cidade:
          type: string
          example: São Paulo
        estado:
          type: string
          minLength: 2
          maxLength: 2
          example: SP
        endereco:
          type: string
          nullable: true
          example: 'Rua Exemplo, 123'
        ativo:
          type: boolean
          example: true
...
    UsuarioSimples:
      type: object
      properties:
        id:
          type: integer
          example: 1
        nome:
          type: string
          example: João Silva
...
    TipoFalha:
      type: object
      properties:
        id:
          type: integer
          example: 1
        nome:
          type: string
          example: Fibra Rompida
        descricao:
          type: string
          nullable: true
          example: Cabo de fibra óptica rompido
...
security:
  - cookieAuth: []